✨ Feat: borderVertices 접기 구현
